Fix Dialogue and Modal files
